+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
 ## 1.1.0 - 2025-11-30
 ### Added
 - Add new clients ,cpp-restbed-server-deprecated, csharp-functions, java-helidon-client, java-helidon-server, java-wiremock, javascript-apollo-deprecated, jetbrains-http-client, julia-client, julia-server, n4js, php-nextgen, postman-collection, python-pydantic-v1, scala-http4s-server, scala-pekko, scala-sttp4, swift-combine